Main loop refactor + BLE handling
* updated main loop with reading ECG/MPU data and updating BLE data
* main loop now doesn't attempt to send BLE data if no BLE TX buffer
  available
* data ready events for both devices set flag checked in main loop
* update to BSP_ECG_ADS1192 with reading data function
* update to BSP_MPU9150 with older (modified) TWI drivers to work with
  SoftDevice, and have lower overhead (faster execution)
* update to BLE event callback that sets flag if no TX buffer available
  for BLE send operation
* Fix timer drivers to work correctly on interrupt (compare) events

+++ application/hal/hal_timer.h
@@ -46,15 +46,15 @@ void HAL_TIMER_setBitWidth(NRF_TIMER_Type *tInstance, DRV_TIMER_bitWidth_E bitWi
 void HAL_TIMER_setMode(NRF_TIMER_Type *tInstance, DRV_TIMER_mode_E mode);
 
 void HAL_TIMER_clearEvents(NRF_TIMER_Type *tInstance);
-void HAL_TIMER_clearEvent(NRF_TIMER_Type *tInstance, uint8_t channel);
+void HAL_TIMER_clearEvent(NRF_TIMER_Type *tInstance, uint32_t channel);
 void HAL_TIMER_runTask(NRF_TIMER_Type *tInstance, DRV_TIMER_task_E task);
 DRV_TIMER_event_E HAL_TIMER_getEvent(uint8_t channel);
-bool HAL_TIMER_checkEvent(NRF_TIMER_Type *tInstance, uint8_t channel);
+bool HAL_TIMER_checkEvent(NRF_TIMER_Type *tInstance, uint32_t channel);
 bool HAL_TIMER_checkIntEn(NRF_TIMER_Type *tInstance, uint8_t channel);
-void HAL_TIMER_enableInterrupt(NRF_TIMER_Type *tInstance, uint8_t channel);
+void HAL_TIMER_enableInterrupt(NRF_TIMER_Type *tInstance, uint32_t channel);
 void HAL_TIMER_disableInterrupt(NRF_TIMER_Type *tInstance, uint8_t channel);
 void HAL_TIMER_writeCompareValue(NRF_TIMER_Type *tInstance,
-        uint8_t channel,
+        uint32_t channel,
         uint32_t compareValue);
 uint32_t HAL_TIMER_getValue(NRF_TIMER_Type *tInstance, DRV_TIMER_cc_E channel);
